Generic form becomes default
