+++ README.md
@@ -1,3 +1,1 @@
-# DeepString
-
-A simple serialization format consisting only of maps, lists, strings, and nulls.
+# Deep
